CQMatrixDialog, CQSplashWidget, CUpDownSubwidget, DefaultplotDialog and objectdebug ui forms converted from Qt3 to Qt version 4.

+++ copasi/UI/CQMatrixDialog.cpp
@@ -1,9 +1,9 @@
 // Begin CVS Header
 //   $Source: /Volumes/Home/Users/shoops/cvs/copasi_dev/copasi/UI/CQMatrixDialog.cpp,v $
-//   $Revision: 1.4.2.2 $
+//   $Revision: 1.4.2.3 $
 //   $Name:  $
 //   $Author: aekamal $
-//   $Date: 2011/01/24 17:00:14 $
+//   $Date: 2011/01/24 17:17:56 $
 // End CVS Header
 
 // Copyright (C) 2010 by Pedro Mendes, Virginia Tech Intellectual
@@ -50,10 +50,10 @@ void CQMatrixDialog::languageChange()
 
 void CQMatrixDialog::setArray(const CArrayAnnotation *tmp, bool single)
 {
-  int i;
+  size_t i;
 
   mpLabelRow->setText("Rows : " + FROM_UTF8(tmp->getDimensionDescription(0)));
-  int nRows = tmp->size()[0];
+  size_t nRows = tmp->size()[0];
 
   if (!single)
     mpCBRow->insertItem("ALL");
@@ -81,7 +81,7 @@ void CQMatrixDialog::setArray(const CArrayAnnotation *tmp, bool single)
       mpCBColumn->show();
 
       mpLabelColumn->setText("Columns : " + FROM_UTF8(tmp->getDimensionDescription(1)));
-      int nCols = tmp->size()[1];
+      size_t nCols = tmp->size()[1];
 
       if (!single)
         mpCBColumn->insertItem("ALL");
@@ -103,7 +103,7 @@ void CQMatrixDialog::setArray(const CArrayAnnotation *tmp, bool single)
           mpCBDim3->show();
 
           mpLabelDim3->setText("Dimension : " + FROM_UTF8(tmp->getDimensionDescription(2)));
-          int nDims = tmp->size()[2];
+          size_t nDims = tmp->size()[2];
 
           if (!single)
             mpCBDim3->insertItem("ALL");
